+++ osf/apps/osf-ui/src/app/tabs/__tests__/configuration-tab.component.spec.ts
@@ -145,6 +145,7 @@ describe('ConfigurationTabComponent', () => {
         bpPlanningApplicationUrl: '',
         bpMesApplicationUrl: '',
         bpEwmApplicationUrl: '',
+        bpCrmApplicationUrl: '',
         bpAnalyticsApplicationUrl: '',
         bpDataLakeApplicationUrl: '',
         dspSmartfactoryDashboardUrl: '',